Jardesson_Criacao_Tela_Home_Refatoracao Refatoração do projeto Remoção de templates
